# Conflicts:
#	db/db_change.sql

+++ doctor-web/src/main/java/io/terminus/doctor/web/front/warehouseV2/vo/WarehousePigGroupApplyVo.java
@@ -45,7 +45,7 @@ public class WarehousePigGroupApplyVo {
 
     private BigDecimal quantity;
 
-    private Long unitPrice;
+    private BigDecimal unitPrice;
 
-    private Long amount;
+    private BigDecimal amount;
 }
